@@ -21,12 +21,12 @@
     }
   ],
   "require": {
-    "php": ">=7.1",
-    "composer-plugin-api": "^1.0 || ^2.0"
+    "php": ">=7.4",
+    "composer-plugin-api": "^2.0"
   },
   "require-dev": {
-    "composer/composer": "1.6.* || 2.0.*@dev",
-    "composer/semver": "^1 || ^3",
+    "composer/composer": "^2.0",
+    "composer/semver": "^3.0",
     "dealerdirect/phpcodesniffer-composer-installer": "^0.7.0",
     "deployer/deployer": "^6.8",
     "wearerequired/coding-standards": "^1.5"
