@@ -27,9 +27,7 @@ set( 'keep_releases', 3 );
 set( 'wordpress', true );
 set(
 	'bin/wp',
-	function () {
-		return locateBinaryPath( 'wp' );
-	}
+	fn(): string => locateBinaryPath( 'wp' )
 );
 
 // Load options and hosts from inventory.
@@ -39,14 +37,14 @@ inventory( 'deploy.yml' );
 desc( 'Install WordPress translations' );
 task(
 	'wp:install_translations',
-	function () {
+	function (): void {
 		if ( ! get( 'wordpress' ) || ! has( 'wp_languages' ) ) {
 			return;
 		}
 
 		within(
 			'{{release_path}}',
-			function () {
+			function (): void {
 				$wp_languages = implode( ' ', get( 'wp_languages' ) );
 				run( "{{bin/wp}} language core install {$wp_languages} --skip-plugins=wordpress-seo" );
 				run( "{{bin/wp}} language plugin install --all {$wp_languages} --format=csv --skip-plugins=wordpress-seo" );
@@ -59,14 +57,14 @@ task(
 desc( 'Update WordPress translations' );
 task(
 	'wp:update_translations',
-	function () {
+	function (): void {
 		if ( ! get( 'wordpress' ) || ! has( 'wp_languages' ) ) {
 			return;
 		}
 
 		within(
 			'{{release_path}}',
-			function () {
+			function (): void {
 				run( '{{bin/wp}} language core update --quiet' );
 				run( '{{bin/wp}} language plugin update --all --quiet' );
 				run( '{{bin/wp}} language theme update --all --quiet' );
@@ -81,14 +79,14 @@ task( 'wp:translations', [ 'wp:install_translations', 'wp:update_translations' ]
 desc( 'Clear OPcache' );
 task(
 	'wp:opcache_clear',
-	function () {
+	function (): void {
 		if ( ! get( 'wordpress' ) || ! get( 'wp_clear_opcache', false ) ) {
 			return;
 		}
 
 		within(
 			'{{release_path}}',
-			function () {
+			function (): void {
 				$is_installed = test( '{{bin/wp}} plugin is-installed wp-cli-clear-opcache' );
 				if ( ! $is_installed ) {
 					writeln( '<comment>Skipped because wp-cli-clear-opcache is not installed.</comment>' );
@@ -104,7 +102,7 @@ task(
 desc( 'Runs the WordPress database update procedure' );
 task(
 	'wp:upgrade_db',
-	function () {
+	function (): void {
 		if ( ! get( 'wordpress' ) ) {
 			return;
 		}
@@ -112,7 +110,7 @@ task(
 		try {
 			within(
 				'{{release_path}}',
-				function () {
+				function (): void {
 					$is_multisite = test( '{{bin/wp}} core is-installed --network' );
 					run( '{{bin/wp}} core update-db' . ( $is_multisite ? ' --network' : '' ) );
 				}
